refactor: moved function to Fixture
